vinyl: add graceful shutdown
Let's stop all vinyl internal fibers and threads.

In case of scheduler it looks like we revert the commit e463128e4bd4
("vinyl: cancel reader and writer threads on shutdown") so we can again
have delay on shutdown in 'vinyl/replica_quota.test'. I guess we should
not.

At the time of the commit deferring deletes was the default behavior and
there is a secondary index in the test space. The deferred deletes
involve TX thread communication and at moment of stopping scheduler
worker threads the TX event loop was not running. This could result in
worker threads hanging on stop. In this patch we stop worker threads in
shutdown phase while TX event loop is active.

We delete part of the test for #3412 as now we finish fibers that may
use the latch. Also we restore destroying the latch.

Part of #8423

NO_CHANGELOG=internal
NO_DOC=internal