box: support omitting arguments on error creation
We are going keep existing format messages for most of existing error
codes as mentioned in more details in the previous commit. But sometimes
we want error constructor argument to be formatted and to be added to
payload differently. One example is `ER_TUPLE_FOUND` we want tuples to
be formatted as string as before in formatted string but want to add
them as `TUPLE`.

For this purpose let's use NULL payload field name. In this case error
constructor argument will be used for formatted string but will not be
added to the error payload.

Follow up #9109

NO_CHANGELOG=internal
NO_DOC=internal