core: test we can free trigger its callback
If we can free trigger in its callback we can keep memory management
simple. For example the trigger can be part of some object structure and
the object can be freed during trigger callback execution.

We already support it. Just add a test.

NO_DOC=add a test
NO_CHANGELOG=add a test