core: add fiber_wait_dead
We already have such logic in the code. Let's factor it out to a function.
We can use it on shutdown to wait for some system fiber shutdown even
if it can be joined from some third fiber destroying object that have
system fiber as worker.

NO_CHANGELOG=internal
NO_DOC=internal