ci: add workflow to check downgrade versions
Tarantool has hardcoded list of versions it can downgrade to. This list
should consist of all the released versions less than Tarantool version.
This workflow helps to make sure we update the list before release.

When run on the commit with release tag it check the list and fail
if it misses some released version less than current. In this case
we are supposed to update downgrade list (with required downgrade code)
and update the release tag.

Closes #8319

NO_TEST=ci
NO_CHANGELOG=ci
NO_DOC=ci