shutdown: shutdown tasks worker fiber
As this fiber is made system in the commit bf620650c429 ("box: finish
client fibers on shutdown") we don not need the existing protection from
cancelling. So first remove it. Now shutdown the fiber using cancel.

We also need changes to the tasks itself. The thing is tasks still
can be scheduled after tasks worker fiber shutdown when event loop is
not running. So the tasks should be able to work without event loop.

NO_CHANGELOG=internal
NO_DOC=internal